requirements.txt
# Binance API
python-binance>=1.0.19

# Veri İşleme
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0

# Teknik Analiz
ta>=0.10.2

# WebSocket ve Async
websocket-client>=1.6.0
aiohttp>=3.8.0

# Konfigürasyon
python-dotenv>=1.0.0
pyyaml>=6.0

# Loglama
loguru>=0.7.0

# Test
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
